pyproject.toml
[project]
name = "projetodio-fastapi"
version = "0.1.0"
description = ""
authors = [
    {name = "Gustavo Pereira",email = "gdnp@academico.ufpb.br"}
]
readme = "README.md"
requires-python = ">=3.13,<4.0"
dependencies = [
    "fastapi[standard] (>=0.116.1,<0.117.0)",
    "sqlalchemy (>=2.0.42,<3.0.0)",
    "alembic (>=1.16.4,<2.0.0)",
    "asyncpg (>=0.30.0,<0.31.0)",
    "pydantic-settings (>=2.10.1,<3.0.0)",
    "psycopg2-binary (>=2.9.10,<3.0.0)"
]

[tool.poetry]
packages = [{include = "projetodio_fastapi", from = "src"}]


[tool.poetry.group.dev.dependencies]
ruff = "^0.12.7"
pytest = "^8.4.1"
httpx = "^0.28.1"
taskipy = "^1.14.1"
pytest-asyncio = "^1.1.0"
aiosqlite = "^0.21.0"

[tool.ruff]
line-length = 79
exclude = [".venv", "venv", ".git", "alembic"]
[tool.ruff.lint]
preview = true
select = [
    "E",
    "F",
    "I",
    "W",
    "PL",
    "PT"]
ignore = ["E501"]

[tool.ruff.format]
quote-style = "single"

[tool.taskipy.tasks]
lint = "ruff check"
pre_format = "ruff check --fix"
format = "ruff format"

[tool.pytest.ini_options]
pythonpath = [
  "."
]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
